Add roles validation

+++ src/pages/MembersPage.js
@@ -1,16 +1,20 @@
 import React, { Component } from 'react';
+import { Modal } from 'reactstrap';
 import Preloader from '../components/Preloader';
 import MembersTable from '../components/MembersTable';
 import Firebase from '../services/Firebase';
 import { Button } from '../UI/Buttons';
 import { addCache, loadCache } from '../utils/cache';
-import Modal from '../UI/Modal';
+import ModalContent from '../UI/ModalContent';
 import FormModal from '../components/FormModal';
 import inputsChangeHandler from '../utils/inputsChangeHandler';
 import { defaultRegisterData } from '../utils/defaultInputsData';
 import DataModal from '../components/DataModal';
 import { membersInputs } from '../utils/inputs';
-import validation from '../utils/validation';
+import { validation } from '../utils/validation';
+import { stringToDate, dateToString } from '../utils/convertDate';
+import AuthContext from '../context';
+import { DangerAlert } from '../UI/Alerts';
 
 export default class MembersPage extends Component {
   constructor() {
@@ -23,6 +27,7 @@ export default class MembersPage extends Component {
       isEditMode: false,
       isDetailMode: false,
       isFormValid: false,
+      showAlert: false,
     };
     this.db = new Firebase();
   }
@@ -53,7 +58,15 @@ export default class MembersPage extends Component {
   };
 
   onModalOpen = () => {
-    this.setState({
+    const {
+      user: { role },
+    } = this.context;
+    if (role === 'MENTOR') {
+      return this.setState({
+        showAlert: true,
+      });
+    }
+    return this.setState({
       showModal: true,
     });
   };
@@ -71,38 +84,49 @@ export default class MembersPage extends Component {
   onFormChange = (e) => {
     const { value, id } = e.target;
     this.setState(({ registerData }) => {
-      const updatedRegisterData = inputsChangeHandler(value, id, registerData);
-      const validatedInputs = { ...updatedRegisterData };
-      delete validatedInputs.id; // delete id of objects as it's should not be validate
+      const updated = inputsChangeHandler(value, id, registerData);
+      const validatedInputs = { ...updated };
       const isFormValid = validation(validatedInputs, membersInputs);
       return {
-        registerData: updatedRegisterData,
+        registerData: updated,
         isFormValid,
       };
     });
   };
 
   onAddNewMember = async (member) => {
-    await this.db.addNewUser(member);
+    const { birthDate, startDate } = member;
+    const newMember = { ...member, birthDate: stringToDate(birthDate), startDate: stringToDate(startDate) };
+    await this.db.addNewUser(newMember);
     const result = await this.getMembersData();
     this.onModalClose();
     return result;
   };
 
   onEditMemberModalOpen = (userId) => {
+    const {
+      user: { role },
+    } = this.context;
+    if (role === 'MENTOR') {
+      return this.setState({
+        showAlert: true,
+      });
+    }
     const { members } = this.state;
     const editedUser = members.find(({ id }) => id === userId);
+    const { birthDate, startDate } = editedUser;
     this.onModalOpen();
-    this.setState({
-      registerData: { ...editedUser },
+    return this.setState({
+      registerData: { ...editedUser, birthDate: dateToString(birthDate), startDate: dateToString(startDate) },
       isEditMode: true,
       isFormValid: true,
     });
   };
 
-  onSubmitEditUser = async () => {
-    const { registerData } = this.state;
-    await this.db.editUserData(registerData);
+  onSubmitEditUser = async (member) => {
+    const { birthDate, startDate } = member;
+    const newMember = { ...member, birthDate: stringToDate(birthDate), startDate: stringToDate(startDate) };
+    await this.db.editUserData(newMember);
     const result = await this.getMembersData();
     this.onModalClose();
     return result;
@@ -111,45 +135,66 @@ export default class MembersPage extends Component {
   onMemberDataOpen = (userId) => {
     const { members } = this.state;
     const editedUser = members.find(({ id }) => id === userId);
-    this.onModalOpen();
     this.setState({
+      showModal: true,
       registerData: { ...editedUser },
       isDetailMode: true,
     });
   };
 
   onUserDelete = async (userId) => {
+    const {
+      user: { role },
+    } = this.context;
+    if (role === 'MENTOR') {
+      return this.setState({
+        showAlert: true,
+      });
+    }
     await this.db.deleteUser(userId);
     const result = await this.getMembersData();
     return result;
   };
 
+  onAlertClose = () => {
+    this.setState({
+      showAlert: false,
+    });
+  };
+
+  onSubmit = () => {
+    const { isEditMode, registerData } = this.state;
+    return isEditMode ? this.onSubmitEditUser(registerData) : this.onAddNewMember(registerData);
+  };
+
   render() {
-    const { members, isLoaded, showModal, registerData, isEditMode, isDetailMode, isFormValid } = this.state;
+    const { members, isLoaded, showModal, registerData, isEditMode, isDetailMode, isFormValid, showAlert } = this.state;
     const btnStyles = { marginBottom: '1rem' };
     const modalHeader =
       isEditMode || isDetailMode ? <h3>{`${registerData.name}'s details:`}</h3> : <h3>Add new user:</h3>;
     return (
       <div className='table-wrapper'>
-        <Modal
-          showModal={showModal}
-          isEditMode={isEditMode}
-          isDetailMode={isDetailMode}
-          onModalClose={this.onModalClose}
-          isFormValid={isFormValid}
-          onSubmit={() => (isEditMode ? this.onSubmitEditUser(registerData) : this.onAddNewMember(registerData))}
-        >
-          {isDetailMode ? (
-            <DataModal header={modalHeader} data={registerData} inputFields={membersInputs} />
-          ) : (
-            <FormModal
-              inputs={membersInputs}
-              data={registerData}
-              onFormChange={this.onFormChange}
-              isEditMode={isEditMode}
-              modalHeader={modalHeader}
-            />
-          )}
+        <Modal isOpen={showModal} toggle={this.onModalClose}>
+          <ModalContent
+            showModal={showModal}
+            isEditMode={isEditMode}
+            isDetailMode={isDetailMode}
+            onModalClose={this.onModalClose}
+            isFormValid={isFormValid}
+            onSubmit={this.onSubmit}
+          >
+            {isDetailMode ? (
+              <DataModal header={modalHeader} data={registerData} inputFields={membersInputs} />
+            ) : (
+              <FormModal
+                inputs={membersInputs}
+                data={registerData}
+                onFormChange={this.onFormChange}
+                isEditMode={isEditMode}
+                modalHeader={modalHeader}
+              />
+            )}
+          </ModalContent>
         </Modal>
         {isLoaded ? (
           <>
@@ -162,6 +207,9 @@ export default class MembersPage extends Component {
               onMemberDataOpen={this.onMemberDataOpen}
               onUserDelete={this.onUserDelete}
             />
+            <DangerAlert isOpen={showAlert} toggle={this.onAlertClose}>
+              This feature available only for admin
+            </DangerAlert>
           </>
         ) : (
           <Preloader />
@@ -170,3 +218,5 @@ export default class MembersPage extends Component {
     );
   }
 }
+
+MembersPage.contextType = AuthContext;
